tests/wpt/web-platform-tests/workers/constructors/SharedWorker/URLMismatchError.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Workers: SharedWorker - throw URLMismatchError</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

  test(function() {
    var worker = new SharedWorker('shared-worker.js', 'name');

    assert_throws("URLMismatchError", function() {
      new SharedWorker('some-other-url.js', 'name');
    });

  }, "Create SharedWorker with different URLs but same name");

</script>
