Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/constructors/SharedWorker/URLMismatchError-expected.txt

CONSOLE MESSAGE: line 17: ReferenceError: Can't find variable: SharedWorker

FAIL Web Workers: SharedWorker - same name, different URL ReferenceError: Can't find variable: SharedWorker
